Refactor join module: Split 713-line file into focused modules
Split crates/executor/src/select/join/mod.rs into three focused modules
for improved maintainability and code organization:

- hash_join.rs: Hash join implementation with all related tests
- nested_loop.rs: All nested loop join variants (INNER, LEFT, RIGHT, FULL, CROSS)
- mod.rs: Public API and join orchestration logic

Benefits:
- Each module is ~250 lines (vs 713 in original)
- Better code discoverability
- Easier testing and maintenance
- Zero runtime impact (code organization only)
- No breaking changes to public API

All 313 executor library tests pass, including all 5 hash join tests
in the new module structure.

Closes #786
